+++ scripts/install-pdfchecker.sh
@@ -6,11 +6,13 @@
 
 set -o pipefail -o errexit -o nounset -o xtrace
 
+BASE_URL=https://cdn.datalogics.com/www/pdflibinfo/grzqordg/che-lin64-MfeaURpvtQXzGKjY
+BASE_URL=https://www.datalogics.com/pdflibinfo/grzqordg/che-lin64-MfeaURpvtQXzGKjY
 DOWNLOADED_ZIP_FILENAME=PDF-CHECKER-Lin64.zip
 INSTALL_DIR_PATH=${1:-$PWD/PDF_Checker}
 export TMPDIR=$(mktemp -d /tmp/pdfchecker.XXXXXX)
 
-wget --quiet --no-check-certificate https://www.datalogics.com/pdflibinfo/grzqordg/che-lin64-MfeaURpvtQXzGKjY/$DOWNLOADED_ZIP_FILENAME
+wget --quiet --no-check-certificate $BASE_URL/$DOWNLOADED_ZIP_FILENAME
 unzip $DOWNLOADED_ZIP_FILENAME
 rm $DOWNLOADED_ZIP_FILENAME